+++ src/docchatai/app/config.py
@@ -82,21 +82,48 @@ class AppConfig:
 
 @unique
 class ChatVar(str, Enum):
-    def __new__(cls, value):
+    def __new__(cls, value, sys_arg_index: int = -1):
         obj = str.__new__(cls, [value])
         obj._value_ = value
+        obj.__sys_arg_index = sys_arg_index
         return obj
-    MODEL_PROVIDER = 'chat_model_provider'
-    MODEL = 'chat_model'
-    FILE = 'chat_file'
+
+    @property
+    def sys_arg_index(self) -> int:
+        return self.__sys_arg_index
+
+    MODEL_PROVIDER = ('chat_model_provider', 1)
+    MODEL = ('chat_model', 2)
+    FILE = ('chat_file', 3)
     TEMPLATE = 'chat_template'
     REQUEST = 'chat_request'
 
 class ChatConfig:
+    @staticmethod
+    def from_sys_args(app_config: AppConfig = AppConfig()) -> 'ChatConfig':
+        args = {}
+        for e in ChatVar:
+            en = ChatVar(e)
+            if 0 < en.sys_arg_index < len(argv):
+                args[str(en.value)] = argv[en.sys_arg_index]
+        return ChatConfig(app_config, args)
+
+    @staticmethod
+    def from_dict(app_config: AppConfig = AppConfig(), values: dict[str, str] = None) -> 'ChatConfig':
+        return ChatConfig(app_config, values)
+
     def __init__(self, app_config: AppConfig = AppConfig(), values: dict[str, str] = None):
         super().__init__()
         self.__app_config = app_config
         self.__values = {} if values is None else values
+
+    def to_dict(self) -> dict[str, str]:
+        return {
+            str(ChatVar.MODEL_PROVIDER.value): self.chat_model_provider,
+            str(ChatVar.MODEL.value): self.chat_model_name,
+            str(ChatVar.FILE.value): self.chat_file,
+            str(ChatVar.TEMPLATE.value): self.chat_template
+        }
 
     @property
     def app_config(self) -> AppConfig:
@@ -105,27 +132,22 @@ class ChatConfig:
     @property
     def chat_model_provider(self) -> str:
         return self._get_val_key_case_insensitive(
-            ChatVar.MODEL_PROVIDER, argv[1] if len(argv) > 1 else self.app_config.default_chat_model_name)
+            ChatVar.MODEL_PROVIDER, self.app_config.default_chat_model_name)
 
     @property
     def chat_model_name(self) -> str:
-        return self._get_val_key_case_insensitive(
-            ChatVar.MODEL, argv[2] if len(argv) > 2 else self.app_config.default_chat_model_name)
+        return self._get_val_key_case_insensitive(ChatVar.MODEL, self.app_config.default_chat_model_name)
 
     @property
     def chat_file(self) -> str:
-        return self._get_val_key_case_insensitive(ChatVar.FILE, argv[3] if len(argv) > 3 else None)
+        return self._get_val_key_case_insensitive(ChatVar.FILE, None)
 
     @property
     def chat_template(self) -> str:
         return self._get_val_key_case_insensitive(ChatVar.TEMPLATE, self.app_config.default_chat_template)
 
-    @property
-    def chat_request(self) -> str:
-        return self._get_val_key_case_insensitive(ChatVar.REQUEST, None)
-
     def _get_val_key_case_insensitive(self, key: ChatVar, default: str or None) -> str or None:
-        key = key.value
+        key = str(key.value)
         value = self.__values.get(key, os.environ.get(key, None))
         if value is None:
             key = key.upper()
@@ -135,4 +157,4 @@ class ChatConfig:
     def __str__(self):
         return (f'RunConfig: model={self.chat_model_name}, '
                 f'max_result_per_query={self.app_config.max_results_per_query}, '
-                f'chat_file={self.chat_file}, chat_request={self.chat_request}')
+                f'chat_file={self.chat_file}, chat_template={self.chat_template}')